+++ chapter4/ezytutors/tutor-nodb/src/models.rs
@@ -4,8 +4,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct Course {
-    pub tutor_id: usize,
-    pub course_id: Option<usize>,
+    pub tutor_id: i32,
+    pub course_id: Option<i32>,
     pub course_name: String,
     pub posted_time: Option<NaiveDateTime>,
 }
